0016: 코드 공통화
