+++ src/main/kotlin/com/smparkworld/discord/bot/CommandHandler.kt
@@ -1,8 +1,6 @@
 package com.smparkworld.discord.bot
 
-import com.smparkworld.discord.base.StringCode
-import com.smparkworld.discord.base.StringsParser
-import net.dv8tion.jda.api.EmbedBuilder
+import com.smparkworld.discord.extensions.sendUnknownExceptionMessage
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent
 import net.dv8tion.jda.api.events.interaction.component.ButtonInteractionEvent
 import net.dv8tion.jda.api.events.interaction.component.EntitySelectInteractionEvent
@@ -15,11 +13,7 @@ abstract class CommandHandler {
             handle(command, event)
         } catch (e: Exception) {
             e.printStackTrace()
-
-            val message = EmbedBuilder()
-                .setDescription(StringsParser.getString(StringCode.UNKNOWN_EXCEPTION))
-                .build()
-            event.replyEmbeds(message).queue()
+            event.sendUnknownExceptionMessage()
         }
     }
     protected abstract fun handle(command: String, event: SlashCommandInteractionEvent)
