+++ src/main/kotlin/com/smparkworld/discord/bot/bee/commands/BeeMusicPlayLeaveCommandHandler.kt
@@ -1,13 +1,14 @@
 package com.smparkworld.discord.bot.bee.commands
 
 import com.smparkworld.discord.base.StringCode
-import com.smparkworld.discord.base.StringsParser
 import com.smparkworld.discord.base.StringsParser.getString
 import com.smparkworld.discord.bot.CommandHandler
 import com.smparkworld.discord.bot.bee.commands.player.MusicManagerMediator
+import com.smparkworld.discord.extensions.checkVoiceChannelValidation
+import com.smparkworld.discord.extensions.sendEmbedsMessage
+import com.smparkworld.discord.extensions.sendUnknownExceptionMessage
 import com.smparkworld.discord.usecase.GetVoiceChannelByEventAuthorUseCase
 import net.dv8tion.jda.api.EmbedBuilder
-import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel
 import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent
 import kotlin.random.Random
 
@@ -17,14 +18,11 @@ class BeeMusicPlayLeaveCommandHandler(
 ) : CommandHandler() {
 
     override fun handle(command: String, event: SlashCommandInteractionEvent) {
-        checkVoiceChannelValidation(event) { _ ->
+        checkVoiceChannelValidation(event, result = getVoiceChannelByEventAuthor(event)) { _ ->
 
             val guild = event.guild
             if (guild == null) {
-                val message = EmbedBuilder()
-                    .setDescription(getString(StringCode.UNKNOWN_EXCEPTION))
-                    .build()
-                event.replyEmbeds(message).queue()
+                event.sendUnknownExceptionMessage()
                 return@checkVoiceChannelValidation
             }
 
@@ -42,30 +40,7 @@ class BeeMusicPlayLeaveCommandHandler(
             val message = EmbedBuilder()
                 .setDescription(msg)
                 .build()
-            event.replyEmbeds(message).queue()
-        }
-    }
-
-    private fun checkVoiceChannelValidation(
-        event: SlashCommandInteractionEvent,
-        perform: (channel: VoiceChannel) -> Unit
-    ) {
-        when (val result = getVoiceChannelByEventAuthor(event)) {
-            is GetVoiceChannelByEventAuthorUseCase.Result.Success -> {
-                perform.invoke(result.voiceChannel)
-            }
-            is GetVoiceChannelByEventAuthorUseCase.Result.NotInVoiceChannel -> {
-                val message = EmbedBuilder()
-                    .setDescription(getString(StringCode.ABSENT_COMMAND_AUTHOR))
-                    .build()
-                event.replyEmbeds(message).queue()
-            }
-            is GetVoiceChannelByEventAuthorUseCase.Result.Error -> {
-                val message = EmbedBuilder()
-                    .setDescription(getString(StringCode.UNKNOWN_EXCEPTION))
-                    .build()
-                event.replyEmbeds(message).queue()
-            }
+            event.sendEmbedsMessage(message)
         }
     }
 }
