+++ src/main/kotlin/com/smparkworld/discord/bot/valorant/commands/ValorantTeamCommandHandler.kt
@@ -3,6 +3,8 @@ package com.smparkworld.discord.bot.valorant.commands
 import com.smparkworld.discord.base.StringCode
 import com.smparkworld.discord.base.StringsParser.getString
 import com.smparkworld.discord.bot.CommandHandler
+import com.smparkworld.discord.extensions.checkAudioChannelValidation
+import com.smparkworld.discord.extensions.sendEmbedsMessage
 import com.smparkworld.discord.usecase.GetVoiceChannelByNameUseCase
 import com.smparkworld.discord.usecase.GetVoiceChannelUsersByEventAuthorUseCase
 import net.dv8tion.jda.api.EmbedBuilder
@@ -22,7 +24,7 @@ class ValorantTeamCommandHandler(
 ) : CommandHandler() {
 
     override fun handle(command: String, event: SlashCommandInteractionEvent) {
-        checkAudioChannelValidation(event) { members ->
+        checkAudioChannelValidation(event, result = getVoiceChannelUsersByMember(event)) { members ->
 
             val ignores: List<User> = obtainIgnoredUsers(event)
             val players: List<Member> = obtainPlayers(members, ignores)
@@ -95,30 +97,7 @@ class ValorantTeamCommandHandler(
             .setTitle(getString(StringCode.VAL_TEAM_FINISH_TITLE))
             .setDescription(getString(StringCode.VAL_TEAM_FINISH_DESC, finishAuthor.toString()))
             .build()
-        event.replyEmbeds(message).queue()
-    }
-
-    private fun checkAudioChannelValidation(
-        event: SlashCommandInteractionEvent,
-        perform: (members: List<Member>) -> Unit
-    ) {
-        when (val result = getVoiceChannelUsersByMember(event)) {
-            is GetVoiceChannelUsersByEventAuthorUseCase.Result.Success -> {
-                perform.invoke(result.members)
-            }
-            is GetVoiceChannelUsersByEventAuthorUseCase.Result.NotInVoiceChannel -> {
-                val message = EmbedBuilder()
-                    .setDescription(getString(StringCode.ABSENT_COMMAND_AUTHOR))
-                    .build()
-                event.replyEmbeds(message).queue()
-            }
-            is GetVoiceChannelUsersByEventAuthorUseCase.Result.Error -> {
-                val message = EmbedBuilder()
-                    .setDescription(getString(StringCode.UNKNOWN_EXCEPTION))
-                    .build()
-                event.replyEmbeds(message).queue()
-            }
-        }
+        event.sendEmbedsMessage(message)
     }
 
     private fun checkPlayersValidation(
@@ -131,13 +110,13 @@ class ValorantTeamCommandHandler(
                 val message = EmbedBuilder()
                     .setDescription(getString(StringCode.VAL_TEAM_CANDIDATE_NEED_TO_MORE))
                     .build()
-                event.replyEmbeds(message).queue()
+                event.sendEmbedsMessage(message)
             }
             (players.size > 10) -> {
                 val message = EmbedBuilder()
                     .setDescription(getString(StringCode.VAL_TEAM_CANDIDATE_TOO_MUCH))
                     .build()
-                event.replyEmbeds(message).queue()
+                event.sendEmbedsMessage(message)
             }
             else -> perform.invoke()
         }
